+++ src/device/orbita_foc.rs
@@ -1,4 +1,4 @@
-//! Orbita Serial SimpleFOC register (protocol v1)
+//! Orbita 3DoF Serial SimpleFOC register (protocol v1)
 
 use crate::device::*;
 
@@ -38,7 +38,8 @@ reg_read_only!(firmware_version, 6, u8);
 reg_read_write!(id, 7, u8);
 
 reg_write_only!(system_check, 8, u8);
-reg_read_only!(motors_drivers_states, 9, u8);
+
+reg_read_only!(motors_drivers_states, 159, u8);
 
 reg_read_write!(voltage_limit, 10, f32);
 // reg_read_write!(intensity_limit, 14, f32);
@@ -70,6 +71,11 @@ reg_read_only!(present_position, 71, DiskValue::<f32>);
 reg_read_only!(top_present_position, 71, f32);
 reg_read_only!(middle_present_position, 75, f32);
 reg_read_only!(bottom_present_position, 79, f32);
+
+reg_read_only!(top_present_hall, 160, f32);
+reg_read_only!(middle_present_hall, 164, f32);
+reg_read_only!(bottom_present_hall, 168, f32);
+
 // reg_read_only!(present_speed, 83, DiskValue::<f32>);
 // reg_read_only!(top_present_speed, 83, f32);
 // reg_read_only!(middle_present_speed, 87, f32);
@@ -93,6 +99,10 @@ reg_read_only!(bottom_present_position, 79, f32);
 // reg_read_only!(imu_gyro_y, 135, f32);
 // reg_read_only!(imu_gyro_z, 139, f32);
 // reg_read_only!(imu_temperature, 143, f32);
+
+reg_read_only!(top_current_phase_a, 147, f32);
+reg_read_only!(top_current_phase_b, 151, f32);
+reg_read_only!(top_dc_current, 155, f32);
 
 impl<T: PartialEq> PartialEq for DiskValue<T> {
     fn eq(&self, other: &Self) -> bool {