+++ src/protocol/v1.rs
@@ -8,6 +8,7 @@ use super::{CommunicationErrorKind, Packet};
 const BROADCAST_ID: u8 = 254;
 const BROADCAST_RESPONSE_ID: u8 = 253;
 
+#[derive(Debug)]
 pub struct PacketV1;
 impl Packet for PacketV1 {
     const HEADER_SIZE: usize = 4;
@@ -94,6 +95,7 @@ impl Packet for PacketV1 {
     }
 }
 
+#[derive(Debug)]
 struct InstructionPacketV1 {
     id: u8,
     instruction: InstructionKindV1,
@@ -126,6 +128,7 @@ impl InstructionPacket<PacketV1> for InstructionPacketV1 {
     }
 }
 
+#[derive(Debug)]
 struct StatusPacketV1 {
     id: u8,
     #[allow(dead_code)]
@@ -145,6 +148,11 @@ impl StatusPacket<PacketV1> for StatusPacketV1 {
         let read_crc = *data.last().unwrap();
         let computed_crc = crc(&data[2..data.len() - 1]);
         if read_crc != computed_crc {
+            println!(
+                "read crc: {}, computed crc: {} data: {:?}",
+                read_crc, computed_crc, data
+            );
+
             return Err(Box::new(CommunicationErrorKind::ChecksumError));
         }
 
@@ -195,7 +203,6 @@ pub enum DynamixelErrorV1 {
 impl DynamixelErrorV1 {
     fn from_byte(error: u8) -> Vec<Self> {
         (0..7)
-            .into_iter()
             .filter(|i| error & (1 << i) != 0)
             .map(|i| DynamixelErrorV1::from_bit(i).unwrap())
             .collect()
@@ -235,6 +242,7 @@ impl InstructionKindV1 {
     }
 }
 
+#[derive(Debug)]
 pub struct V1;
 impl Protocol<PacketV1> for V1 {
     fn new() -> Self {
@@ -302,11 +310,11 @@ mod tests {
 
     #[test]
     fn create_write_packet() {
-        let p = PacketV1::write_packet(10, 24, &vec![1]);
+        let p = PacketV1::write_packet(10, 24, &[1]);
         let bytes = p.to_bytes();
         assert_eq!(bytes, [255, 255, 10, 4, 3, 24, 1, 213]);
 
-        let p = PacketV1::write_packet(0xFE, 0x03, &vec![1]);
+        let p = PacketV1::write_packet(0xFE, 0x03, &[1]);
         let bytes = p.to_bytes();
         assert_eq!(bytes, [0xFF, 0xFF, 0xFE, 0x04, 0x03, 0x03, 0x01, 0xF6]);
     }
@@ -323,7 +331,7 @@ mod tests {
 
     #[test]
     fn create_sync_write_packet() {
-        let p = PacketV1::sync_write_packet(&[11, 12], 30, &vec![vec![0x0, 0x0], vec![0xA, 0x14]]);
+        let p = PacketV1::sync_write_packet(&[11, 12], 30, &[vec![0x0, 0x0], vec![0xA, 0x14]]);
         let bytes = p.to_bytes();
         assert_eq!(
             bytes,