+++ Changelog.md
@@ -1,3 +1,12 @@
+## Version 0.5.0
+
+- Add an post delay option to the read and write method.
+- Add dxl XM motor device
+
+## Version 0.4.0
+
+- Add support for orbita-2dof-foc device.
+
 ### Version 0.3.1
 
 - Patch torque limit conversion.